local-business-mcp/requirements.txt
# Core requirements
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.15
geoalchemy2>=0.13.0
psycopg2-binary>=2.9.6
alembic>=1.10.4

# API Clients
googlemaps>=4.10.0
facebook-sdk>=3.1.0
requests>=2.28.2
aiohttp>=3.8.4

# Data processing
pandas>=2.0.0
numpy>=1.24.3
scikit-learn>=1.2.2

# Export utilities
openpyxl>=3.1.2
jinja2>=3.1.2
pdfkit>=1.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
click>=8.1.3
rich>=13.3.5
geopy>=2.3.0
folium>=0.14.0
